+++ doomsday/engine/api/sys_musd.h
@@ -53,14 +53,14 @@ typedef struct musinterface_generic_s {
 // Driver interface for playing MUS music.
 typedef struct musinterface_mus_s {
     musinterface_generic_t gen;
-    void           *(*SongBuffer) (int length);
+    void           *(*SongBuffer) (size_t length);
     int             (*Play) (int looped);
 } musinterface_mus_t;
 
 // Driver interface for playing non-MUS music.
 typedef struct musinterface_ext_s {
     musinterface_generic_t gen;
-    void           *(*SongBuffer) (int length);
+    void           *(*SongBuffer) (size_t length);
     int             (*PlayFile) (const char *filename, int looped);
     int             (*PlayBuffer) (int looped);
 } musinterface_ext_t;
