+++ doomsday/engine/portable/include/sv_def.h
@@ -61,11 +61,11 @@ void            Sv_Ticker(void);
 int             Sv_Latency(byte cmdtime);
 void            Sv_Kick(int who);
 void            Sv_GetInfo(serverinfo_t *info);
-int             Sv_InfoToString(serverinfo_t *info, ddstring_t * msg);
+size_t          Sv_InfoToString(serverinfo_t *info, ddstring_t *msg);
 boolean         Sv_StringToInfo(const char *valuePair, serverinfo_t *info);
 int             Sv_GetNumPlayers(void);
 int             Sv_GetNumConnected(void);
-void            Sv_PlaceMobj(struct mobj_s* mo, float x, float y, float z, boolean onFloor);
+void            Sv_PlaceMobj(struct mobj_s *mo, float x, float y, float z, boolean onFloor);
 
 boolean         Sv_CheckBandwidth(int playerNumber);
